HEC-437: Split command.rb (411 lines) into validation, lifecycle, and dispatch modules
Extract pre/postcondition DSL into command/validation.rb (Hecks::Command::Validation)
and event construction/persistence into command/dispatch.rb (Hecks::Command::Dispatch).
command.rb drops from 411 to 259 total lines (~160 code lines after doc headers).
